Fix bug with prop count randomization
Prop count for an item is selected in scrambleRow by using the randInt method. This method was passed the min and max value in from the scrambler, but the function can not return the maximum value. It returns values in the range [min, max). This change adds 1 to the maximum when calling RandInt to get values including the maximum.