Add option to balance property count
Adds a configuration option that balances the property count on
randomized items. This setting will make the randomizer pick prop
count for each item from a vanilla item that is up to 10 levels above
the item being randomized.

Updates the default config to have vanilla unique drop rate and
turn on balanced randomization.