Add PreserveVanilla option
Adds PreserveVanilla Randomizer option.  This will clone all of the rows given to the randomizer and modify props on the clones.  Clones are all named with original name + an "R".  All say "An Evil Force" when viewed in game because patchstring.tbl has not been updated with the new names.